+++ src/storage/storageIOBars.ts
@@ -20,21 +20,30 @@
 
 import GObject from 'gi://GObject';
 
+import Config from '../config.js';
 import BarsBase, { BarProps } from '../bars.js';
 import { StorageIO } from './storageMonitor.js';
 
 type StorageIOBarsParams = BarProps & {
-    colors?: string[]
+    /* empty */
 };
 
 export default GObject.registerClass(
 class StorageIOBars extends BarsBase {
     constructor(params: StorageIOBarsParams) {
-        //TODO: Make these configurable
-        if(params.colors === undefined)
-            params.colors = ['rgb(29,172,214)', 'rgb(214,29,29)'];
-        
         super(params);
+        
+        Config.connect(this, 'changed::storage-header-io-bars-color1', this.setStyle.bind(this));
+        Config.connect(this, 'changed::storage-header-io-bars-color2', this.setStyle.bind(this));
+    }
+    
+    setStyle() {
+        super.setStyle();
+        
+        this.colors = [
+            Config.get_string('storage-header-io-bars-color1') ?? 'rgba(29,172,214,1.0)',
+            Config.get_string('storage-header-io-bars-color2') ?? 'rgba(214,29,29,1.0)'
+        ];
     }
     
     setUsage(usage: StorageIO[]|null) {